# java-services/chat-service/src/main/java/org/marre/chat/dto/ChatResponse.java
package org.marre.chat.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.Data;

@Data
public class ChatResponse {
    private String response;
    private String model;
    private Boolean success;

    private TokenUsage usage;
    private String sessionId;

    @Data
    public static class TokenUsage {
        @JsonProperty("prompt_tokens")
        private Integer promptTokens;

        @JsonProperty("completion_tokens")
        private Integer completionTokens;

        @JsonProperty("total_tokens")
        private Integer totalTokens;
    }
}